When client is shutting down some invocations were left hanging.
A test added for that case.
To fix, invocation service and connection manager heavily refactored.

To increase readability. Invocation service and connection manager interfaces are introduced.
Other services access these two over interface. Methods that are private to these classes are not
exposed on interface.
== Connection Manager ==
None of the connection manager methods returns channel. Old getOrConnect was creating
channel in every access. getOrTriggerConnect is introduced instead of it. If connection
is available, we will not create any channel. Also implementation is close to reference
implementation(java client) this way.

== Invocation Service ==
Usage of channels are removed. Main reason was that it is hard to close channels
properly. Only channel left is to handle responses(to avoid stalling response thread).

Accesses to maos like responseWaitings etc. Guarded by RWMutex.
== Invocation ==
Response channels for error and normal response on invocation is merged to one. Reasons
1. deadlocks caused by double completion can be handled from one entry point
2. Performance: Create one channel per invocation instead of two

Complex timeout logic is simplified.

== Others ==
Address and member interfaces are derived from Stringer

core.Address is used instead of protocol.Address, (interface instead of struct)
on connection manager and invocation service
Further cleanup can be done in all protocol.Address usages.